+++ src/cpp11.cpp
@@ -43,5 +43,4 @@ static const R_CallMethodDef CallEntries[] = {
 extern "C" void R_init_svglite(DllInfo* dll){
   R_registerRoutines(dll, NULL, CallEntries, NULL, NULL);
   R_useDynamicSymbols(dll, FALSE);
-  R_forceSymbols(dll, TRUE);
 }